+++ app/src/main/java/com/example/clase6/fragment/FragmentForm.kt
@@ -1,0 +1,26 @@
+package com.example.clase6.fragment
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.clase6.databinding.FragmentFormBinding
+
+class FragmentForm : Fragment() {
+
+    private lateinit var binding: FragmentFormBinding
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentFormBinding.inflate(inflater)
+        //binding.lifeCicleOwner= this
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+    }
+
+}
