+++ app/src/main/res/layout/fragment_form.xml
@@ -1,0 +1,111 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <androidx.appcompat.widget.Toolbar
+        android:id="@+id/toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="?attr/actionBarSize"
+        android:background="?attr/colorPrimary"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:title="Nueva Cita" />
+
+    <EditText
+        android:id="@+id/etNombreMascota"
+        android:hint="Nombre de la Mascota"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/toolbar"
+        app:theme="@style/AppTheme.InputField"
+        android:layout_width="221dp"
+        android:layout_height="57dp"
+        android:layout_marginStart="108dp"
+        android:layout_marginTop="40dp"
+        android:imeOptions="actionNext"
+        android:inputType="textPersonName"
+        android:maxLength="15"
+        android:padding="16dp"
+        android:textColor="@android:color/black" />
+
+
+    <EditText
+        android:id="@+id/etRazaMascota"
+        android:layout_width="223dp"
+        android:layout_height="53dp"
+        android:layout_marginStart="108dp"
+        android:layout_marginTop="12dp"
+        android:hint="Raza"
+        android:imeOptions="actionNext"
+        android:inputType="textVisiblePassword"
+        android:padding="16dp"
+        android:textColor="@android:color/black"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/etNombreMascota"
+        app:theme="@style/AppTheme.InputField" />
+
+    <EditText
+        android:id="@+id/etNombrePropietario"
+        android:layout_width="223dp"
+        android:layout_height="59dp"
+        android:layout_marginStart="108dp"
+        android:layout_marginTop="8dp"
+        android:hint="Nombre del Propietario"
+        android:imeOptions="actionNext"
+        android:inputType="textPersonName"
+        android:maxLength="30"
+        android:padding="16dp"
+        android:textColor="@android:color/black"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/etRazaMascota"
+        app:theme="@style/AppTheme.InputField" />
+
+    <EditText
+        android:id="@+id/etTelefonoPropietario"
+        android:layout_width="225dp"
+        android:layout_height="56dp"
+        android:layout_marginStart="108dp"
+        android:layout_marginTop="20dp"
+        android:hint="Teléfono"
+        android:imeOptions="actionNext"
+        android:inputType="phone"
+        android:maxLength="10"
+        android:padding="16dp"
+        android:textColor="@android:color/black"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/etNombrePropietario"
+        app:theme="@style/AppTheme.InputField" />
+
+    <Spinner
+        android:id="@+id/spSintomas"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="108dp"
+        android:layout_marginTop="24dp"
+        android:entries="@array/sintomas"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/etTelefonoPropietario" />
+
+    <Button
+        android:id="@+id/btnGuardar"
+        android:layout_width="347dp"
+        android:layout_height="58dp"
+        android:layout_marginStart="16dp"
+        android:layout_marginTop="32dp"
+        android:layout_marginEnd="16dp"
+        android:layout_marginBottom="16dp"
+        android:text="Guardar"
+        android:theme="@style/AppTheme.SaveButton"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/etRazaMascota" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
+
+
+
+
