+++ app/src/main/res/navigation/nav_graph.xml
@@ -15,7 +15,7 @@
             app:destination="@id/fragmentB" />
         <action
             android:id="@+id/action_fragmentA_to_fragmentCardView"
-            app:destination="@id/fragmentCardView" />
+            app:destination="@id/fragmentForm" />
     </fragment>
     <fragment
         android:id="@+id/fragmentB"
@@ -24,8 +24,8 @@
         tools:layout = "@layout/fragment_b">
     </fragment>
     <fragment
-        android:id="@+id/fragmentCardView"
-        android:name="com.example.clase6.fragment.FragmentCardView"
-        android:label="FragmentCardView"
-        tools:layout = "@layout/fragment_card_view"/>
+        android:id="@+id/fragmentForm"
+        android:name="com.example.clase6.fragment.FragmentForm"
+        android:label="fragmentForm"
+        tools:layout = "@layout/fragment_form"/>
 </navigation>
