+++ app/src/main/res/values/colors.xml
@@ -3,4 +3,5 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="purple">#D736F4</color>
+    <color name="saveButtonColor">#F05861</color>
 </resources>
