+++ app/src/main/res/values/styles.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <style name="AppTheme.SaveButton" parent="Widget.AppCompat.Button">
+        <item name="android:backgroundTint">@color/saveButtonColor</item> <item name="android:radius">8dp</item> <item name="android:padding">16dp</item> <item name="android:textSize">16sp</item> <item name="android:textStyle">bold</item> </style>
+    <style name="AppTheme.InputField" parent="Widget.AppCompat.EditText">
+        <item name="android:background">@color/white</item> <item name="android:textColor">@android:color/black</item> <item name="android:padding">16dp</item> <item name="android:radius">8dp</item> <item name="android:strokeWidth">1</item> <item name="android:strokeColor">@android:color/black</item> </style>
+
+    <style name="corner">
+        <item name="android:layout_width">221dp</item>
+        <item name="android:layout_height">57dp</item>
+        <item name="android:layout_marginStart">108dp</item>
+        <item name="android:layout_marginTop">40dp</item>
+        <item name="android:imeOptions">actionNext</item>
+        <item name="android:inputType">textPersonName</item>
+        <item name="android:maxLength">15</item>
+        <item name="android:padding">16dp</item>
+        <item name="android:textColor">@android:color/black</item>
+    </style>
+
+</resources>
